Configurable paycheck at configurable progress
