+++ Assets/Paycheck.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48b4517f612564b88a8e1be06043a4d4
+timeCreated: 1438515628
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
