+++ Assets/Scripts/Paycheck.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4bb19f659fd9a42dd9a4b0129ffb31d4
-timeCreated: 1438508772
+guid: 75053213a4ca742eca91d1fd294e7b2f
+timeCreated: 1438513960
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
