+++ Assets/Scripts/Thing.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 25960d58944484262b9f6fdf8fab5bb3
+folderAsset: yes
+timeCreated: 1438514225
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
