+++ Assets/Scripts/Thing/ThingController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 027f04ffd27ae4d91963571084eb47ac
+timeCreated: 1438514227
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
