+++ Assets/Scripts/Thing/ThingHighlighter.cs
