+++ Assets/Scripts/Thing/ThingState.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: f2bc65742cd4b442387473ddd357d092
+timeCreated: 1438514236
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
