--- Assets/Scripts/ThingController.cs.meta
@@ -1,15 +1,0 @@
-fileFormatVersion: 2
-guid: bbbe3d2295f76ed4dbbca048530b4d2f
-timeCreated: 1437332008
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences:
-  - prefabDoneSparks: {fileID: 19871470, guid: 5c818c726b1244d4f97e904c01ffc7f9, type: 2}
-  - prefabProgressText: {fileID: 10289820, guid: 41dd0047565a1478c9552df798bc73ab,
-      type: 2}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
