+++ Assets/Thing.prefab
@@ -83,7 +83,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 157514}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbbe3d2295f76ed4dbbca048530b4d2f, type: 3}
+  m_Script: {fileID: 11500000, guid: 027f04ffd27ae4d91963571084eb47ac, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   prefabDoneSparks: {fileID: 19871470, guid: 5c818c726b1244d4f97e904c01ffc7f9, type: 2}
@@ -99,7 +99,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 157514}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4bb19f659fd9a42dd9a4b0129ffb31d4, type: 3}
+  m_Script: {fileID: 11500000, guid: 85cc189c6c1d7498da595625917fd700, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   isHighlighted: 0
